loader, dialog and post of carta porte

+++ src/app/modules/carta-porte/pages/complete/complete.component.ts
@@ -1,4 +1,5 @@
 import {
+  AfterViewInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
@@ -11,26 +12,31 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Subscription, catchError, map, tap } from 'rxjs';
 
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 
 import { CartaPorteService } from '../../services/cartaporte.service';
 import { CatalogsService } from '../../services/catalogs.service';
 import { ErrorsService } from 'src/app/modules/core/services/errors.service';
 import { LocalStorageService } from '../../services/localStorage.service';
 
-import { Driver } from '../../interfaces/driver.interface';
-import { Location, originLocation } from '../../interfaces/locations.interface';
-import { Vehicle } from '../../interfaces/vehicle.interface';
-import { PostCartaPorte } from '../../interfaces/cartaporte.interface';
+import {
+  Driver,
+  Location,
+  originLocation,
+  Vehicle,
+  PostCartaPorte,
+  PostDates,
+} from '../../interfaces';
+import { LoadingDialogComponent } from '../../components/loading-dialog/loading-dialog.component';
 
 @Component({
   selector: 'app-complete',
   templateUrl: './complete.component.html',
   styleUrls: ['./complete.component.scss'],
 })
-export class CompleteComponent implements OnInit, OnDestroy {
+export class CompleteComponent implements OnInit, AfterViewInit, OnDestroy {
   private cartaPorteService: CartaPorteService = inject(CartaPorteService);
   private catalogsService: CatalogsService = inject(CatalogsService);
-  private cdr: ChangeDetectorRef = inject(ChangeDetectorRef);
   private errorsService: ErrorsService = inject(ErrorsService);
   private localStorageService: LocalStorageService = inject(LocalStorageService);
 
@@ -38,36 +44,45 @@ export class CompleteComponent implements OnInit, OnDestroy {
   private route: ActivatedRoute = inject(ActivatedRoute);
   private router: Router = inject(Router);
   private snackBar: MatSnackBar = inject(MatSnackBar);
+  private cdr: ChangeDetectorRef = inject(ChangeDetectorRef);
+  private dialog: MatDialog = inject(MatDialog);
 
   private routeSubscription!: Subscription;
-
-  private locationID?: number;
   private cartaPorteID?: number;
+  private dates?: PostDates;
 
   public orderNumber: string = '';
   public originLocation: Location = originLocation;
   public destinyLocation?: Location;
-  public locationLoading: boolean = false;
+  public locationLoading: boolean = true;
+  public currentDate: Date = new Date();
+
   public currentVehicle?: Vehicle;
   public currentDriver?: Driver;
-  public drivers?: Driver[];
-  public vehicles?: Vehicle[];
-  public currentDate: Date = new Date();
+
+  public drivers: Driver[] = [];
+  public vehicles: Vehicle[] = [];
 
   public form: FormGroup = this.fb.group({
     vehicle: ['', [Validators.required]],
     driver: ['', [Validators.required]],
-    date: ['', [Validators.required]],
-    hour: ['', [Validators.required]],
+    exitDate: ['', [Validators.required]],
+    exitHour: ['', [Validators.required]],
+    arriveDate: ['', [Validators.required]],
+    arriveHour: ['', [Validators.required]],
   });
 
   ngOnInit() {
     this.routeSubscription = this.route.params.subscribe((params) => {
       this.orderNumber = params['orderNumber'];
     });
+  }
 
+  ngAfterViewInit(): void {
     this.form.get('vehicle')?.disable();
     this.form.get('driver')?.disable();
+    this.form.get('arriveHour')?.disable();
+    this.form.get('arriveDate')?.disable();
 
     this.setLocation();
     this.getVehicles();
@@ -89,14 +104,43 @@ export class CompleteComponent implements OnInit, OnDestroy {
   }
 
   public setDates(): void {
-    const exitDate: Date = this.form.get('date')?.value;
+    const exitDate: Date = this.form.get('exitDate')?.value;
+    const hour = this.form.get('exitHour')?.value;
 
-    if (!exitDate) return;
+    if (!exitDate || !hour) return;
 
     const arriveDate: Date = this.getThreeWorkDaysAfter(exitDate);
 
-    const formatExitDate: string = this.formatDate(exitDate);
-    const formatArriveDate: string = this.formatDate(arriveDate);
+    const formatExitDate: string = this.formatDate(exitDate, hour);
+    const formatArriveDate: string = this.formatDate(arriveDate, hour);
+
+    this.dates = {
+      exitDate: formatExitDate,
+      arriveDate: formatArriveDate,
+    };
+  }
+
+  public setDay(): void {
+    const exitDate: Date = this.form.get('exitDate')?.value;
+    if (!exitDate) return;
+
+    const arriveDate: Date = this.getThreeWorkDaysAfter(exitDate);
+    const plainArriveDate: string = `${arriveDate.getDate()}/${
+      arriveDate.getMonth() + 1
+    }/${arriveDate.getFullYear()}`;
+
+    this.form.get('arriveDate')?.setValue(plainArriveDate);
+
+    this.setDates();
+  }
+
+  public setHour(): void {
+    const exitHour = this.form.get('exitHour')?.value;
+    if (!exitHour) return;
+
+    this.form.get('arriveHour')?.setValue(exitHour);
+
+    this.setDates();
   }
 
   public setVehicle(): void {
@@ -124,14 +168,24 @@ export class CompleteComponent implements OnInit, OnDestroy {
       return;
     }
 
-    this.localStorageService.deleteKeyFromHistory(this.orderNumber);
-    this.successSnackBar();
-    this.router.navigateByUrl('list');
-    console.log(this.buildCartaPorte());
+    const cartaPorte = this.buildCartaPorte();
+    if (!cartaPorte) return;
+
+    const loadingDialog = this.loadingDialog();
+    this.cartaPorteService.postCartaPorte(cartaPorte).subscribe((result) => {
+      loadingDialog.close();
+      if (result === true) {
+        this.localStorageService.deleteKeyFromHistory(this.orderNumber);
+        this.successSnackBar();
+        this.router.navigateByUrl('list');
+        return;
+      }
+
+      this.errorSnakBar(`Error al generar Carta Porte ${this.orderNumber}`);
+    });
   }
 
   private setLocation(): void {
-    this.locationLoading = true;
     this.cartaPorteService
       .getList()
       .pipe(
@@ -152,7 +206,6 @@ export class CompleteComponent implements OnInit, OnDestroy {
         })
       )
       .subscribe((locationID) => {
-        this.locationID = locationID;
         this.getLocation(locationID);
       });
   }
@@ -215,19 +268,16 @@ export class CompleteComponent implements OnInit, OnDestroy {
     this.snackBar.open(message, btnLabel, config);
   }
 
-  private formatDate(date: Date): string {
+  private formatDate(date: Date, hour: string): string {
     const year = date.getFullYear();
     const month = ('0' + (date.getMonth() + 1)).slice(-2);
     const day = ('0' + date.getDate()).slice(-2);
-    const hour = ('0' + date.getHours()).slice(-2);
-    const minutes = ('0' + date.getMinutes()).slice(-2);
-    const seconds = ('0' + date.getSeconds()).slice(-2);
 
-    return `${year}-${month}-${day}T${hour}:${minutes}:${seconds}`;
+    return `${year}-${month}-${day}T${hour}:00`;
   }
 
   private getThreeWorkDaysAfter(date: Date): Date {
-    let afterDate: number | Date = date.setDate(date.getDate() + 3);
+    let afterDate: number | Date = new Date().setDate(date.getDate() + 3);
     afterDate = new Date(afterDate);
 
     if (afterDate.getDay() === 0) {
@@ -239,16 +289,24 @@ export class CompleteComponent implements OnInit, OnDestroy {
   }
 
   private buildCartaPorte(): PostCartaPorte | undefined {
-    if (!this.currentDriver || !this.currentVehicle || !this.cartaPorteID) return;
+    if (!this.currentDriver || !this.currentVehicle || !this.cartaPorteID || !this.dates)
+      return;
 
     const cartaPorte: PostCartaPorte = {
       idSAT_CartaPorte: this.cartaPorteID,
       idSAT_Driver: this.currentDriver.idSAT_Driver,
       idSAT_vehicle: this.currentVehicle.idSAT_vehicle,
-      FechaHoraSalida: '9:00',
-      FechaHoraLlegada: '10:00',
+      FechaHoraSalida: this.dates.exitDate,
+      FechaHoraLlegada: this.dates.arriveDate,
     };
 
     return cartaPorte;
+  }
+
+  private loadingDialog(): MatDialogRef<LoadingDialogComponent> {
+    return this.dialog.open(LoadingDialogComponent, {
+      data: this.orderNumber,
+      disableClose: true,
+    });
   }
 }
